cleaned up filesystem
Cleaned up filesystem, folder within folder
