--- moviefindr/src/main/java/com/moviefindr/MoviefindrApplication.java
@@ -1,13 +1,0 @@
-package com.moviefindr;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MoviefindrApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MoviefindrApplication.class, args);
-	}
-
-}
