--- moviefindr/src/main/resources/templates/error.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <title>HomeScreen</title>
-    <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
-    <link rel="stylesheet" href="assets/css/Navigation-Clean.css">
-    <link rel="stylesheet" href="assets/css/Navigation-with-Button.css">
-    <link rel="stylesheet" href="assets/css/styles.css">
-</head>
-
-<body style="color: rgb(251,251,252);background: #a5a7ad;">
-    You have reached a wrong page which does not exist please go back home
-</body>
-
-</html>
