--- moviefindr/src/main/resources/templates/genre.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <title>Select Your Genres</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-</head>
-
-<body style="color: rgb(251,251,252);background: #a5a7ad;">
-    <nav class="navbar navbar-light navbar-expand-md navigation-clean" style="background: rgb(90,92,98);">
-        <div class="container"><a class="navbar-brand" href="#" style="color: rgb(248,240,240);">Movie Findr</a>
-            <button data-toggle="collapse" class="navbar-toggler" data-target="#navcol-1">
-                <span class="sr-only">Toggle navigation></span>
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div style="width: 980px;height: 30px;text-align: right;padding-top: 3px;">
-                <small style="color: rgb(247,245,245);font-size: 15px;">Find Something to Watch</small>
-            </div>
-        <div class="collapse navbar-collapse" id="navcol-1">
-            <ul class="nav navbar-nav ml-auto"></ul>
-        </div>
-        </div>
-    </nav>
-    <div style="height: 650px;text-align: center;margin-top: 75px;">
-        <h1 style="margin-bottom: 25px;">Select the Genre(s):</h1>
-        <div style="height: 40px;margin-bottom: 10px;padding-top: 0;margin-top: 10px;">
-            <!-- The three dropdowns are below -->
-            <div class="dropdown d-inline-block" style="border-style: none;">
-                <button id="dropdownBtn1" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-expanded="false" type="button" style="width: 180px;background: rgb(126,130,134);text-align: center;border-style: none;">Select a Genre</button>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item" href="#">Horror</a>
-                    <a class="dropdown-item" href="#">Drama</a>
-                    <a class="dropdown-item" href="#">Action</a>
-                    <a class="dropdown-item" href="#">Romance</a>
-                    <a class="dropdown-item" href="#">Comedy</a>
-                </div>
-            </div>
-        </div>
-        <div style="height: 40px;margin-bottom: 10px;">
-            <div class="dropdown d-inline-block">
-                <button id="dropdownBtn2" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-expanded="false" type="button" style="width: 180px;background: rgb(126,130,134);border-style: none;">Select a Genre</button>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item" href="#">Horror</a>
-                    <a class="dropdown-item" href="#">Drama</a>
-                    <a class="dropdown-item" href="#">Action</a>
-                    <a class="dropdown-item" href="#">Romance</a>
-                    <a class="dropdown-item" href="#">Comedy</a>
-                </div>
-            </div>
-        </div>
-        <div style="height: 40px;margin-bottom: 10px;">
-            <div class="dropdown d-inline-block">
-                <button id="dropdownBtn3" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-expanded="false" type="button" style="width: 180px;background: rgb(126,130,134);border-style: none;">Select a Genre</button>
-                <div class="dropdown-menu">
-                    <a class="dropdown-item" href="#">Horror</a>
-                    <a class="dropdown-item" href="#">Drama</a>
-                    <a class="dropdown-item" href="#">Action</a>
-                    <a class="dropdown-item" href="#">Romance</a>
-                    <a class="dropdown-item" href="#">Comedy</a>
-                </div>
-            </div>
-        </div>
-        <!-- Button choose genres and move to the voting page -->
-    <a th:href="@{/voteForm}" id="findMovie" class="btn btn-primary text-center" type="button" style="background: rgb(90,92,98);padding-top: 6px;margin-top: 20px;border-width: 0px;border-style: none;width: 180px;">Find Movies!</a></div>
-
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js" integrity="sha384-b/U6ypiBEHpOf/4+1nzFpr53nxSS+GLCkfwBdFNTxtclqqenISfwAzpKaMNFNmj4" crossorigin="anonymous"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/js/bootstrap.min.js" integrity="sha384-h0AbiXch4ZDo7tp9hKZ4TsHbi047NrKGLO3SEJAg45jXxnGIfYzk4Si90RDIqNm1" crossorigin="anonymous"></script>
-</body>
-
-</html>
