--- moviefindr/src/main/resources/templates/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8" xmlns:th="https://www.thymeLeaf.org">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <title>Home</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-</head>
-
-<body style="color: rgb(251,251,252);background: #a5a7ad;">
-    <nav class="navbar navbar-light navbar-expand-md navigation-clean" style="background: rgb(90,92,98);">
-        <div class="container"><a class="navbar-brand" href="#" style="color: rgb(248,240,240);">Movie Findr</a>
-            <button id="redirectBtn" data-toggle="collapse" class="navbar-toggler" data-target="#navcol-1">
-                <span class="sr-only">Toggle navigation</span>
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div style="width: 980px;height: 30px;text-align: right;padding-top: 3px;">
-                <small style="color: rgb(247,245,245);font-size: 15px;padding-top: 0;">Find Something to Watch</small>
-            </div>
-        <div class="collapse navbar-collapse" id="navcol-1">
-            <ul class="nav navbar-nav ml-auto"></ul>
-        </div>
-        </div>
-    </nav>
-    <div style="height: 650px;text-align: center;">
-        <h1 style="margin-top: 150px;">Welcome to Movie Findr!</h1>
-        <!-- Button to navigate to the genre page -->
-        <a th:href="@{/genreForm}" class="btn btn-primary text-center" type="button" style="background: rgb(90,92,98);padding-top: 6px;margin-top: 50px;border-width: 0px;border-style: none;">Find Something to Watch</a>
-    </div>
-</body>
-
-</html>
