--- moviefindr/src/test/java/com/moviefindr/MoviefindrApplicationTests.java
@@ -1,13 +1,0 @@
-package com.moviefindr;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MoviefindrApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
