+++ src/main/resources/templates/genre.html
@@ -63,7 +63,7 @@
             </div>
         </div>
         <!-- Button choose genres and move to the voting page -->
-    <button id="findMovie" class="btn btn-primary text-center" type="button" style="background: rgb(90,92,98);padding-top: 6px;margin-top: 20px;border-width: 0px;border-style: none;width: 180px;">Find Movies!</button></div>
+    <a th:href="@{/voteForm}" id="findMovie" class="btn btn-primary text-center" type="button" style="background: rgb(90,92,98);padding-top: 6px;margin-top: 20px;border-width: 0px;border-style: none;width: 180px;">Find Movies!</a></div>
 
     <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js" integrity="sha384-b/U6ypiBEHpOf/4+1nzFpr53nxSS+GLCkfwBdFNTxtclqqenISfwAzpKaMNFNmj4" crossorigin="anonymous"></script>
