+++ src/main/resources/templates/vote.html
@@ -56,10 +56,10 @@
             </table>
         </div>
         <!-- Button for moving to movie selection page -->
-        <button id="selectionBtn" class="btn btn-primary" type="button"
+        <a th:href="@{/selectedForm}" id="selectionBtn" class="btn btn-primary" type="button"
                 style="background: rgb(90,92,98);border-style: none;width: 178px;height: 43px;">
             Show my Selections
-        </button>
+        </a>
     </div>
 </body>
